packet: Refactor DNS handling
Remove the `zone_id` knob. We can easily calculate this value and
have no reason to bother the user with it.

Remove the `aws_creds_path` knob under `dns.route53`. AFAICT the only
case where this knob is useful is if I want to store Terraform state
in S3 on AWS account A *and* I want to use Route 53 on AWS account B.
To me it seems like a rather esoteric use case and therefore I think
it's better to remove this knob for simplicity's sake.

Convert the `dns` configuration block to top-level knobs. With the
`zone_id` and `aws_creds_path` knobs removed, the `route53` block is
now empty. Since the `manual` DNS options has no knobs and since
`dns.zone` is provider-agnostic anyway, I think it is simpler to have
just two top-level DNS-related knobs: `dns_zone` and `dns_provider`,
both simple strings.

Move DNS-related code outside of the Packet Terraform module for a
better separation of concerns and clearer module boundaries. Instead
of calculating the required DNS records inside the Packet module,
return the *controller IPs* as an output to be consumed by the various
DNS modules. This makes it easier to add new DNS providers and opens
the way to supporting multiple DNS providers on platforms other than
Packet. The "manual" DNS option is now a standalone Terraform module
which doesn't create any resources and instead just returns outputs
to be consumed by lokoctl.