terraform: add reusable worker and controller modules
This commit adds 3 Terraform modules, which aims to eventually
deduplicate Ignition configuration between all the platforms, by
providing extendable base configuration.

For now, only Tinkerbell platform will be consuming those modules.
Eventually, we should migrate all platforms to use it, however this is
not trivial task, as we lack tests on this level, so it must be done
carefully.

Terraform modules are added in new `terraform-modules` directory, which
for the time being is nested inside `lokomotive-kubernetes` directory,
as it would require non-trivial changes on the Go level, to extract
extra directory to the user assets. Either that or all Terraform code
should be moved inside this directory and then `lokomotive-kubernetes`
should be removed, as it is no longer meaningful after merging 2
repositories.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>