Refactor Terraform executor
- Define a new type terraform.ExecutionStep which allows describing
  arbitrary Terraform operations flexibly.
- Allow passing one or more ExecutionSteps to the Terraform executor.

This makes `pkg/terraform` more generic and self-contained. Callers
can define an arbitrary series of execution steps and apply them to
an Executor in one go. The Executor then takes care of sequentially
executing the steps. Arbitrary logic can be included with a step
in the PreExecutionHook parameter. This allows callers to optionally
perform preliminary operations before running a specific step.