refactor: add new method to interface to get values template
This commit moves the helm values.yaml template to a separate file
`template.go` for each provider.

In this case, since the only provider is azure, it is moved to
`azure/tempplate.go`.

This allows for multiple templates, each specific to the provider as
each provider will have provider specific BackupStorageLocation and
VolumeSnapshotLocation.

Signed-off-by: Imran Pochi <imran@kinvolk.io>